move callback tests into separate file
also fix syntax errors
